start client flow
