--- src/client/client_flow.rs
@@ -1,13 +1,0 @@
-// 1. Relay Descriptor Lookup
-// 2. Handshake Fee
-// 3. Circuit build
-// 4. Test Bandwidth
-// 5. Init Payments Ledger
-// 6. Client Bandwidth Watcher
-// 7. Circuit Kill. Repeat
-
-async fn client_flow() {}
-
-async fn pick_relays() {
-    // rpc call to lookup descriptor
-}
