+++ src/relay/mod.rs
@@ -1,5 +1,3 @@
 mod start_relay_flow;
-mod select_relay_algo;
 
 pub use start_relay_flow::{start_relay_flow};
-pub use select_relay_algo::{simple_relay_selection_algo};
